+++ src/plenoptic/tools/signal.py
@@ -376,7 +376,7 @@ def add_noise(img: Tensor, noise_mse: Union[float, List[float]]) -> Tensor:
 
     """
     noise_mse = torch.as_tensor(
-        noise_mse, dtype=torch.float32, device=img.device
+        noise_mse, dtype=img.dtype, device=img.device
     ).unsqueeze(0)
     noise_mse = noise_mse.view(noise_mse.nelement(), 1, 1, 1)
     noise = 200 * torch.randn(
